Implemeting getAllMeals.

+++ src/main/java/com/techreturners/teama/healthyfood/api/repository/MealPlanRepository.java
@@ -13,11 +13,11 @@ import java.util.List;
 @Repository
 public interface MealPlanRepository extends CrudRepository<MealPlan, Long> {
     @Query(value = "SELECT * FROM mealplan where userid = :userid ", nativeQuery = true)
-    List<MealPlan> getAllMealPlan(@Param(value = "userid") int userid);
+    List<MealPlan> getAllMealPlan(@Param(value = "userid") Long userid);
 
     @Modifying
     @Query(value = "INSERT INTO mealplan (mealid, userid, date) values(:mealid, :userid, :date) ", nativeQuery = true)
-    List<MealPlan> createMealPlan(@Param(value = "mealid") int mealid,
-                                  @Param(value = "userid") int userid,
+    List<MealPlan> createMealPlan(@Param(value = "mealid") Long mealid,
+                                  @Param(value = "userid") Long userid,
                                   @Param(value = "date") LocalDateTime date);
 }
